# [opt](arm) Optimize the BlockBloomFilter::bucket_find on ARM platforms using NEON instructions.